ソースコードをsrcディレクトリにまとめて、出力先もまとめました
