+++ .gitignore
@@ -56,5 +56,6 @@ keyについて.txt
 
 
 python/venv
+python/image/training/output/image
 ios/Flutter/DartDefines.xcconfig
 ios/Runner/GoogleService-Info.plist
