+++ python/image/training/output/log/class_labels.txt
@@ -1,0 +1,5 @@
+ヒトカゲ: 0
+フシギソウ: 1
+フシギダネ: 2
+フシギバナ: 3
+リザード: 4
