+++ python/image/training/output/log/training_log.csv
@@ -1,0 +1,4 @@
+epoch,accuracy,loss,val_accuracy,val_loss
+0,0.6833333373069763,0.7841087579727173,0.6000000238418579,1.0548354387283325
+1,1.0,0.0031196353957057,0.6000000238418579,2.5125269889831543
+2,1.0,0.0003869272186420858,0.5,3.820620059967041
